pyproject.toml
[tool.poetry]
name = "xardas-next"
version = "0.1.0"
description = ""
authors = ["Luis Michaelis <private@lmichaelis.de>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
discord-py-interactions = "^3.0.2"
pydantic = "^1.8.2"
PyYAML = "^5.4.1"
httpx = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
